Added a dynamic route to handle user channels. Fixed Twitch Auth.
