+++ src/routes/+page.server.ts
@@ -1,0 +1,6 @@
+import { redirect } from '@sveltejs/kit';
+import type { PageServerLoad } from './$types';
+
+export const load = (async () => {
+    return redirect(301, `/buddha`)
+}) satisfies PageServerLoad;
